ref(query hooks): move to each folder
